requirements.txt
crispy-bootstrap4==2023.1
Django==4.2.6
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple==3.8
